Fix some backend issues
After some testing I found that the meeting creation functions weren't working because of some SQL statement issues (apparently we can't use variable replacement for table names, but because the source of the table name isn't user based I'm ok with going around that part since SQL injection is unlikely to be an issue there) so updating the table creation statement.

There were also some issues with the HubPool declaration that were fixed. I've been able to confirm that the websocket hubs are created and added to the map successfully at least.