Update Vue function to handle async calls
Fetch is an async function and I need to handle for that using await, and also adding async to the method function so I'm allowed to use the await call. This appears to get me to the point that I can connect to the websocket, next is figuring out how to actually update the tables properly with meeting participant info.