It works for one person per meeting
I've at least gotten it to the point that the stack updates in real time properly but only for the meeting host. I think the issue is that the websocket connection isn't properly getting broadcast messages pushed.... that or it's some other sort of issue I don't understand. I'll need to go back and look at the chat app I'm basing this on.