Move "unregister" actions from hub to client handler
I mistakenly thought that the hub.unregister channel handler was the place to push out updates to clients in the event of a client disconnect. Turns out this should happen in the clients handler in the writePump of all places. The updates here add a removal of a user from the speaker queue stack db table if their client disconnects, and then a broadcast of the current stack out to the rest of the users.

Updating some logging statements along with this for clarity.

Updating the environment variable name to disable websocket origin checking to be slightly more explicit.