Enable Dockerhub build/push via GitHub Actions
Moving to a release system that will trigger Dockerhub builds on release so we have a bit more separation and more appropriate pipeline to go from local changes to live site updates. Will need to migrate the DO App Platform settings to point to Dockerhub after this and a build is confirmed.

In addition to Dockerhub builds I've also implemented gofmt and golangci-lint jobs in the GitHub Actions so we can get some formatting and linting going in this build pipeline.

I also migrated the go.mod/go.sum files into their proper location inside the actual module and updated the settings so that things would work properly for that.