Resolve golangci-lint issues
Mostly just not doing proper error checking on some of my functions. Main confusing thing was the pruner.go for loop. S1000 dictates that I should not use a select with a single use case and should instead just use a "simple channel send/receive" which basically just means if the channel I'm watching gets a message kick something off I guess. I'm really not entirely sure but it compiles and runs and golangci-lint is happy so w/e.