Remove SQL dependencies Pt. 1
First step in removing SQL dependencies and moving to a client side handler for this app. There are a lot of changes here and it's hard to explain everything but the basics are that I'm moving all data stored in a SQL database to having all the information being stored client side instead, and users will send minimal amounts of string data in messages that include the source client requestor ID, the action, and the meetingID, and potential more data down the line but just some basic stuff for now, and that will get pushed out to the client side apps to be handled by the JS application. This commit should handle 90% of what is required to make this work, the last part will be ensuring that any mod requests are coming from the actual mods of the meeting, which will be simple but requires me to define the actions in the JS app first so I know what is needed.