Fix user register broadcast
The user registration message only needs to go to the client that is registering, not all the clients. This change fixes that by sending to the client.send channel instead of the client.hub.broadcast channel.

Additionally updating the UserMessage struct to include a generic "message data" property that will be used to send what ever data is necessary for an action. An example is, for newUser actions, we will send along the name of the user in a nested JSON object that has a single clientName key-value pair. We will also be doing something similar if a user decides to update their name or something similar.